+++ packages/article-label/src/style/index.js
@@ -1,8 +1,12 @@
-import sharedStyles from "./shared";
+import { fonts, fontSizes } from "@times-components/styleguide";
 
 const styles = {
   title: {
-    ...sharedStyles.title
+    fontFamily: fonts.supporting,
+    fontSize: fontSizes.cardMetaMobile,
+    fontWeight: "400",
+    letterSpacing: 1.2,
+    marginBottom: 0
   }
 };