+++ packages/utils/src/index.js
@@ -1,5 +1,7 @@
 export * from "./faketime";
 export * from "./graphql";
 export * from "./strings";
+
+export { default as clean } from "./props";
 export { default as addMissingProtocol } from "./add-missing-protocol";
 export { default as schema } from "./schema.json";