+++ packages/video/src/touchable.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { TouchableOpacity } from "react-native";
+
+const Touchable = props => <TouchableOpacity {...props} />;
+
+export default Touchable;