+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.3"></a>
+## [0.15.3](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.2...@times-components/article-byline@0.15.3) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.15.2"></a>
+## [0.15.2](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.1...@times-components/article-byline@0.15.2) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.15.1"></a>
+## [0.15.1](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.15.0...@times-components/article-byline@0.15.1) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
 <a name="0.15.0"></a>
 # [0.15.0](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.14.22...@times-components/article-byline@0.15.0) (2018-04-17)