+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.23.4"></a>
+## [0.23.4](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.23.3...@times-components/article-summary@0.23.4) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.23.3"></a>
+## [0.23.3](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.23.2...@times-components/article-summary@0.23.3) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.23.2"></a>
+## [0.23.2](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.23.1...@times-components/article-summary@0.23.2) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.23.1"></a>
+## [0.23.1](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.23.0...@times-components/article-summary@0.23.1) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
 <a name="0.23.0"></a>
 # [0.23.0](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.22.20...@times-components/article-summary@0.23.0) (2018-04-17)