+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.11"></a>
+## [0.2.11](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.10...@times-components/related-articles@0.2.11) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.10"></a>
+## [0.2.10](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.9...@times-components/related-articles@0.2.10) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.9"></a>
+## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.8...@times-components/related-articles@0.2.9) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.8"></a>
+## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.7...@times-components/related-articles@0.2.8) (2018-04-20)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.7"></a>
+## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.6...@times-components/related-articles@0.2.7) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.5...@times-components/related-articles@0.2.6) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.5"></a>
+## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.4...@times-components/related-articles@0.2.5) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.3...@times-components/related-articles@0.2.4) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.2...@times-components/related-articles@0.2.3) (2018-04-19)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.1...@times-components/related-articles@0.2.2) (2018-04-18)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.2.1"></a>
 ## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.2.0...@times-components/related-articles@0.2.1) (2018-04-18)