+++ packages/tealium-utils/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.2.2...@times-components/tealium-utils@0.2.3) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.2.1...@times-components/tealium-utils@0.2.2) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.2.0...@times-components/tealium-utils@0.2.1) (2018-04-20)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.1.7...@times-components/tealium-utils@0.2.0) (2018-04-19)
+
+
+### Features
+
+* BREAKING CHANGE: bump version ([#882](https://github.com/newsuk/times-components/issues/882)) ([6a9198d](https://github.com/newsuk/times-components/commit/6a9198d))
+
+
+
+
+<a name="0.1.7"></a>
+## [0.1.7](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.1.6...@times-components/tealium-utils@0.1.7) (2018-04-19)
+
+
+### Bug Fixes
+
+* fixed tealium storybook for android / native ([#883](https://github.com/newsuk/times-components/issues/883)) ([31666fb](https://github.com/newsuk/times-components/commit/31666fb))
+
+
+
+
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.1.5...@times-components/tealium-utils@0.1.6) (2018-04-18)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
 <a name="0.1.5"></a>
 ## [0.1.5](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.1.4...@times-components/tealium-utils@0.1.5) (2018-04-17)